// src/app/com/net/apps/powerplant/client/core/user.ts
/**
 * Swagger sample app
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface User {
    /**
     * Id
     */
    id?: number;
    /**
     * User login
     */
    username?: string;
    /**
     * Uer first name
     */
    firstName?: string;
    /**
     * User last name
     */
    lastName?: string;
    /**
     * User password ???
     */
    password?: string;
    /**
     * User Status
     */
    userStatus?: number;
}